Routes refactor and table MD working with this version.
